extract the ROBOT_MODEL_FORMAT and associated utility methods for usage in orogen components
