+++ src/kdl_parser.cpp
@@ -46,29 +46,6 @@ using namespace KDL;
 
 namespace kdl_parser{
 
-const char* formatNameFromID(int type)
-{
-    switch(type)
-    {
-        case ROBOT_MODEL_AUTO: return "autodetected";
-        case ROBOT_MODEL_URDF: return "URDF";
-        case ROBOT_MODEL_SDF: return "SDF";
-        default:
-            throw std::invalid_argument("invalid model type ID given");
-    }
-}
-
-ROBOT_MODEL_FORMAT guessFormatFromFilename(const std::string& file)
-{
-    if (file.substr(file.size() - 4) == ".sdf")
-        return ROBOT_MODEL_SDF;
-    if (file.substr(file.size() - 5) == ".urdf")
-        return ROBOT_MODEL_URDF;
-    if (file.substr(file.size() - 6) == ".world")
-        return ROBOT_MODEL_SDF;
-    throw std::invalid_argument("cannot guess robot model format for " + file);
-}
-
 /**
  * load kdl tree from sdf xml string
  */
